% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliva.R
\name{auc}
\alias{auc}
\title{Compute area under the curve with respect to ground and increase}
\usage{
auc(
  data,
  saliva_type = "cortisol",
  remove_s0 = FALSE,
  compute_auc_post = FALSE,
  sample_times = NULL
)
}
\arguments{
\item{data}{Saliva data or canonical results.}

\item{saliva_type}{Measurement column name(s).}

\item{remove_s0}{Whether to remove baseline sample S0.}

\item{compute_auc_post}{Whether to add post-baseline AUCi.}

\item{sample_times}{Sampling-time vector or column name.}
}
\value{
Group-level AUC values.
}
\description{
Compute area under the curve with respect to ground and increase
}
