% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{find_non_compliant_samples}
\alias{find_non_compliant_samples}
\title{Identify non-compliant samples}
\usage{
find_non_compliant_samples(data)
}
\arguments{
\item{data}{Canonical results or sample events.}
}
\value{
Rows with failed compliance.
}
\description{
Identify non-compliant samples
}
