% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{find_sampling_anomalies}
\alias{find_sampling_anomalies}
\title{Identify tube and day mismatches}
\usage{
find_sampling_anomalies(data)
}
\arguments{
\item{data}{Canonical results or sample events.}
}
\value{
Rows with sampling anomalies.
}
\description{
Identify tube and day mismatches
}
