% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliva.R
\name{initial_value}
\alias{initial_value}
\title{Return the initial value in each saliva curve}
\usage{
initial_value(data, saliva_type = "cortisol", remove_s0 = FALSE)
}
\arguments{
\item{data}{Saliva data or canonical results.}

\item{saliva_type}{Measurement column name(s).}

\item{remove_s0}{Whether to remove baseline sample S0.}
}
\value{
Group-level initial values.
}
\description{
Return the initial value in each saliva curve
}
