% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliva.R
\name{max_increase}
\alias{max_increase}
\title{Compute maximum increase from the initial sample}
\usage{
max_increase(
  data,
  saliva_type = "cortisol",
  remove_s0 = FALSE,
  percent = FALSE
)
}
\arguments{
\item{data}{Saliva data or canonical results.}

\item{saliva_type}{Measurement column name(s).}

\item{remove_s0}{Whether to remove baseline sample S0.}

\item{percent}{Whether to return the percentage increase.}
}
\value{
Group-level maximum increase.
}
\description{
Compute maximum increase from the initial sample
}
