% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliva.R
\name{max_value}
\alias{max_value}
\title{Compute the maximum measured value in each saliva curve}
\usage{
max_value(data, saliva_type = "cortisol", remove_s0 = FALSE)
}
\arguments{
\item{data}{Saliva data or canonical results.}

\item{saliva_type}{Measurement column name(s).}

\item{remove_s0}{Whether to remove baseline sample S0.}
}
\value{
Group-level maximum values.
}
\description{
Compute the maximum measured value in each saliva curve
}
