% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliva.R
\name{mean_se}
\alias{mean_se}
\title{Compute mean and standard error by sample}
\usage{
mean_se(data, saliva_type = "cortisol", group_cols = NULL, remove_s0 = FALSE)
}
\arguments{
\item{data}{Saliva data or canonical results.}

\item{saliva_type}{Measurement column name(s).}

\item{group_cols}{Grouping columns.}

\item{remove_s0}{Whether to remove baseline sample S0.}
}
\value{
Mean and standard error by sample.
}
\description{
Compute mean and standard error by sample
}
