% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_compliance_overview}
\alias{plot_compliance_overview}
\title{Plot cohort sampling compliance}
\usage{
plot_compliance_overview(
  data,
  by = "sample_position",
  view = c("proportion", "heatmap")
)
}
\arguments{
\item{data}{Canonical results or sample events.}

\item{by}{Summary grouping variable.}

\item{view}{Plot form.}
}
\value{
A ggplot object.
}
\description{
Plot cohort sampling compliance
}
