% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_timing_deviation}
\alias{plot_timing_deviation}
\title{Plot timing deviations}
\usage{
plot_timing_deviation(data, by = "sample_position")
}
\arguments{
\item{data}{Canonical results or sample events.}

\item{by}{Grouping variable.}
}
\value{
A ggplot object.
}
\description{
Plot timing deviations
}
