% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliva.R
\name{slope}
\alias{slope}
\title{Compute the slope between two saliva samples}
\usage{
slope(
  data,
  sample_labels = NULL,
  sample_idx = NULL,
  saliva_type = "cortisol",
  sample_times = NULL
)
}
\arguments{
\item{data}{Saliva data or canonical results.}

\item{sample_labels}{Two sample labels.}

\item{sample_idx}{Two one-based sample indices.}

\item{saliva_type}{Measurement column name(s).}

\item{sample_times}{Sampling-time vector or column name.}
}
\value{
Group-level slope values.
}
\description{
Compute the slope between two saliva samples
}
