% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_inputs.R
\name{summarize_source_audit}
\alias{summarize_source_audit}
\title{Compactly summarize a source audit}
\usage{
summarize_source_audit(source_audit)
}
\arguments{
\item{source_audit}{Source audit returned by \code{read_raw_logs_from_participant_dirs()}.}
}
\description{
Compactly summarize a source audit
}
