% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{carwatch-package}
\alias{carwatch}
\alias{carwatch-package}
\title{Package namespace declarations}
\description{
Reconstruct CARWatch registrations from immutable app logs, review conversion anomalies through an auditable two-pass workflow, and combine sampling-compliance information with saliva biomarkers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/carwatch-tools/carwatch-r}
  \item Report bugs at \url{https://github.com/carwatch-tools/carwatch-r/issues}
}

}
\author{
\strong{Maintainer}: Robert Richer \email{robert.richer@fau.de}

Authors:
\itemize{
  \item Robert Richer \email{robert.richer@fau.de}
}

}
\keyword{internal}
