% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliva.R
\name{compute_features}
\alias{compute_features}
\title{Compute the standard CARWatch saliva response feature set}
\usage{
compute_features(
  data,
  saliva_type = "cortisol",
  group_levels = NULL,
  sample_level = NULL,
  sample_times = NULL,
  slope_pairs = NULL,
  remove_s0 = FALSE
)
}
\arguments{
\item{data}{Saliva data or canonical results.}

\item{saliva_type}{Measurement column name(s).}

\item{group_levels}{Grouping column names.}

\item{sample_level}{Sample-position column name.}

\item{sample_times}{Sampling-time vector or column name.}

\item{slope_pairs}{Sample pairs for slopes.}

\item{remove_s0}{Whether to remove baseline sample S0.}
}
\value{
Group-level response features.
}
\description{
Compute the standard CARWatch saliva response feature set
}
