% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_results.R
\name{write_study_results}
\alias{write_study_results}
\title{Write canonical CARWatch Study Results}
\usage{
write_study_results(data, path)
}
\arguments{
\item{data}{Complete \code{carwatch_results} object.}

\item{path}{Destination CSV path.}
}
\description{
Write canonical CARWatch Study Results
}
